Fixed issue with logic OR operator
I was having trouble getting the right motor to spin. Issuing MRF255 would spin the left motor forward. I noticed your motor == 'L'||'l' which is a logic OR on the statement "motor == 'L'" OR "'l'", the latter equates to true regardless and it therefore continues as if the instruction is for the left motor. Same occurs in a few places so I propagated the change throughout.